checker: add various private API
We do this in the dictionary instead of the checker so that you can skip
using the checker and go straight to the dictionary. Helpful when you
want to run spellcheck off the main-thread.

Additionally give access to the PangoLanguage based on the language code
of the checker. This can be used as well for off-thread processing of
character runs so we get the right breaking positions.