text-buffer-adapter: use GtkTextBufferCommitNotify
This was supposed to get into our 0.4 release but wasn't completed in time.
We can't rely on signal emittion from GtkTextBuffer because there can be
way to many reordering or inconsistent notifications.

This fixes some issues in where we might try to get information from an
iter that is no longer valid.