provider: improve initial language selection
There are various languages which do not follow spellchecking in the same
fashion as libspelling currently supports.

If we detect that, do not try to use a fallback language so aggressively
so that spellchecking starts essentially "disabled".

You can still enable it by selecting the language you want to use from
the popover menu.

Fixes: GNOME/gnome-text-editor#745